feat: reforco & toasts

+++ solutions/FrontAngular/src/app/view/admin/professor/professor-search/professor-search.component.ts
@@ -4,6 +4,7 @@ import { Router, ActivatedRoute } from '@angular/router';
 import { ProfessorService } from 'src/app/service/professor.service';
 import { TipoAcaoValues } from 'src/app/model/tipo-acao';
 import { MessagesService } from 'src/app/service/messages.service';
+import { TdDialogService } from '@covalent/core/dialogs';
 
 @Component({
   selector: 'app-professor-search',
@@ -24,8 +25,10 @@ export class ProfessorSearchComponent implements OnInit {
    */
   constructor(private router: Router,
             private activatedRoute: ActivatedRoute,
-            //private professorService: ProfessorService,
-            private messageService: MessagesService) { 
+            private professorService: ProfessorService,
+            private messageService: MessagesService,
+            private _dialogService: TdDialogService,
+            private _viewContainerRef: ViewContainerRef) { 
   }
 
   /**
@@ -77,13 +80,32 @@ export class ProfessorSearchComponent implements OnInit {
   }
     
   remover(id: number){
-    this.professorService.remover(id).subscribe(dados => {
-      this.messageService.toastSuccess('Professor excluído com sucesso.');
-      this.listar();
-    },
-    (error: any) => {
-      console.log(error.error.message);
-      this.messageService.toastError(error.error.message);
+    this.openRemoverConfirm(id);
+  }
+
+  openRemoverConfirm(id: number): void {
+    this._dialogService.openConfirm({
+      message: 'Tem certeza que deseja excluir esse professor?',
+      disableClose: true, // defaults to false
+      viewContainerRef: this._viewContainerRef, //OPTIONAL
+      title: 'Excluir professor', //OPTIONAL, hides if not provided
+      cancelButton: 'Não', //OPTIONAL, defaults to 'CANCEL'
+      acceptButton: 'Sim', //OPTIONAL, defaults to 'ACCEPT'
+      width: '500px', //OPTIONAL, defaults to 400px
+    }).afterClosed().subscribe((accept: boolean) => {
+      if (accept) {
+        this.professorService.remover(id).subscribe(dados => {
+          this.messageService.toastSuccess('Professor excluído com sucesso.');
+          this.listar();
+        },
+        (error: any) => {
+          console.log(error.error.message);
+          this.messageService.toastError(error.error.message);
+
+        });
+      } else {
+        // DO SOMETHING ELSE
+      }
       
     });
   }
